Switch to mixed precision

+++ app/py4dgeo_app.cpp
@@ -9,14 +9,14 @@ int
 main()
 {
   // Load a test case
-  std::vector<double> data;
+  std::vector<float> data;
   std::ifstream stream("../tests/data/plane_horizontal_t1.xyz");
   std::string line;
   std::size_t points{ 0 };
 
   while (std::getline(stream, line)) {
     std::istringstream s(line);
-    double x;
+    float x;
     for (int i = 0; i < 3; ++i) {
       s >> x;
       data.push_back(x);
